% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelist.r
\name{check.whitelist}
\alias{check.whitelist}
\title{Check whether a call only calls functions or uses variables that satify a whitelist of allowed symbols}
\usage{
check.whitelist(call, wl.funs = NULL, wl.vars = NULL, bl.funs = NULL,
  bl.vars = NULL, funs = find.funs(call), vars = find.variables(call))
}
\arguments{
\item{call}{the call object}

\item{wl.funs}{a character vector of the function names that are allowed (whitelisted). If NULL ignored.}

\item{wl.vars}{a character vector of the variable names that are allowed (whitelisted). If NULL ignored.}

\item{bl.funs}{a character vector of the function names that are forbidden (blacklisted). If NULL ignored.}

\item{bl.vars}{a character vector of the variable names that are forbidden  (blacklisted). If NULL ignored.}

\item{funs}{by default set to all function calls in call.}

\item{vars}{by default set to all variables used in call.}
}
\description{
The function extracts in a nested fashion all names of called functions and used variables
in a call object. It can look up in whitelists whether these symbols are allowed or in
blacklists whether these symbols are forbidden. Whitelists and blacklists are simple
character vectors with the function / variable names.
}
\details{
Returns a list (ok, fb.funs, fb.vars, msg) where ok is FALSE if a forbidden symbol
has been found and otherwise TRUE. fb.funs and fb.vars list the forbidden function calls
or variable names that have been found. msg is a default error message that can be shown
if the call did not pass the check
}

